Move recent file logic out of AppPreferences and into its own class
I've also fixed how it handles loading an entry that is already in the recents
list. Previously it would leave it in the same place, now it moves it to the
most recently used position.

Moving to a separate class makes it easier to unit test. AppPreferences
is difficult to unit test, since there isn't an easy way to mock
java.util.prefs.Preferences.