Cleanup
There was an inconsistency in VCDTraceLoader: the interface TraceLoader implemented
Closable, but within VCDTraceLoader.load, there was also a try-with-resources around
the file (the file was also not an instance variable, thus the close() method did
nothing). Basically load cleaned up after itself. I removed the Closable based class
from TraceLoader. This is still a bit wonky, but at least is a little more consistent
now.
There are a bunch of other unrelated cleanups here too.