Fix issue introduced in 4b44e51d33c8d390ce8b7d4f1f6ba45bc746bfcc
Range selection was broken because the beginning of the selection jumped to the current
cursor position. This was because setAdjustingCursor() was calling setCursorPosition()
as a side effect to send a notification. I made it send the notification directly to
avoid messing up the selection. Using the cursor changed notification is still a bit hacky.