Fix findBugs error
Measuring memory usage has overhead because it invokes the garbage collector.
I was trying to be clever by making that conditional, but that resulted in a dead
store. I measured it and the gc only takes 50ms on my machine, so I turned it
on by default. Moved measurements into a utility class to declutter the loader.