Clean up progress handling
- Break VCDLoader.load into sub functions to make it simpler.
- Move responsibilty for determining when to send progress updates into ProgressInputStream.
Previously it was in VCDLoader.nextToken, which didn't make any sense. Now ProgressInputStream
takes a listener argument. VCDLoader handles this and forwards it to its listener.
- Break LoadFormatException out of WaveformLoader.LoadInterruptException. The latter didn't make sense if there
was a problem in the file. These are also now both derived from IOException, which simplifies the throws clause a bit.
That might be a bit hacky, since it's arguable whether these are logically IOExceptions or not.