Change how search hints are returned
Previously, I stored search hints for the last search as public members
of ExpressionNode. The reason for doing it this was was to avoid an allocation,
which is the only way to really return multiple values from a function. Since
these functions can be called a lot of times during a search, it can potentially
be called a lot of times.

However, this was kind of clunky, and it wasn't clear that there was a performance issue.
Instead, I now allocate an object and pass it as a parameter to evaluate.  If it turns
out that this is a bottleneck, I can probably find a better way to allocate these,
since there are only a small number live at a time.