Change representation of BitVector
Previously, this was an enumeration of BitValues. This made it easy to
set and get individual bits, but there was more overhead for aggregate
operations on the whole vector. Now, it uses two arrays: one with flags
to indicate z/x, and the other that is just the binary value of the
vector.